+++ app/Jobs/Mail/SendFlexibleMail.php
@@ -42,7 +42,7 @@ class SendFlexibleMail implements ShouldQueue
     {
         return [
             (new RateLimited('emails'))
-                ->releaseAfter(config('mail.rate_limit_release_delay', 60)),
+                ->releaseAfter((int) config('mail.rate_limit_release_delay', 60)),
         ];
     }
