+++ app/Jobs/Mail/SendRegistrationConfirmation.php
@@ -40,7 +40,7 @@ class SendRegistrationConfirmation implements ShouldQueue
     {
         return [
             (new RateLimited('emails'))
-                ->releaseAfter(config('mail.rate_limit_release_delay', 60)),
+                ->releaseAfter((int) config('mail.rate_limit_release_delay', 60)),
         ];
     }
